+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [Unreleased]
 
 ### Added
-- Subcommand `transpose dep strip-dev-features` to remove dev-only features as preparation for publishing.
+- Subcommand `transpose features strip-dev-only` to remove dev-only features as preparation for publishing.
 - Arg `--dep-kind` to `propagate-feature` to allow to exclude specific dep kinds.
 
 ## [0.14.0] - 2023-11-07